refactor(Gallery): consolidate components
