+++ assets/scripts/app/WelcomePanel.scss
@@ -98,7 +98,7 @@ $welcome-panel-box-shadow: $medium-box-shadow;
 }
 
 // Component specific tweaks to StreetName
-// Some duped styles from #gallery element
+// Some duped styles from gallery element
 .welcome-panel .street-name {
   @include street-name-mixin($street-name-gallery-size);
